refactor(tests): extract _make_retry / _no_retry to conftest.py
The same two-helper pair was duplicated across four test files
(test_plugin_saves.py, services/saves/_helpers.py, services/test_game_detail.py,
services/test_playtime.py). Identical bodies, no per-file divergence —
prime drift risk if the Retry Protocol ever changes shape.

Consolidated into tests/conftest.py; existing call sites import from
conftest the same way they already import _make_testable_plugin.

Closes #361.
